+++ src/main/kotlin/icu/windea/pls/lang/navigation/ParadoxScriptedVariableChooseByNameContributor.kt
@@ -1,26 +1,33 @@
-package icu.windea.pls.script.navigation
+package icu.windea.pls.lang.navigation
 
-import com.intellij.navigation.*
-import com.intellij.psi.search.*
-import com.intellij.psi.stubs.*
-import com.intellij.util.*
-import com.intellij.util.indexing.*
+import com.intellij.navigation.ChooseByNameContributorEx
+import com.intellij.navigation.NavigationItem
+import com.intellij.psi.search.GlobalSearchScope
+import com.intellij.psi.stubs.StubIndex
+import com.intellij.util.Processor
+import com.intellij.util.indexing.FindSymbolParameters
+import com.intellij.util.indexing.IdFilter
+import icu.windea.pls.PlsFacade
 import icu.windea.pls.lang.index.ParadoxIndexKeys
-import icu.windea.pls.script.psi.*
+import icu.windea.pls.script.psi.ParadoxScriptScriptedVariable
 
 /**
- * 用于在随处搜索（Search Everywhere）中查找对应名字的封装变量。
+ * 用于在 *随处搜索（Search Everywhere）* 中查找封装变量。
  */
 class ParadoxScriptedVariableChooseByNameContributor : ChooseByNameContributorEx {
     //com.intellij.ide.util.gotoByName.JavaModuleNavigationContributor
 
     private val indexKey = ParadoxIndexKeys.ScriptedVariableName
 
+    private fun isEnabled() = PlsFacade.getSettings().navigation.seForScriptedVariables
+
     override fun processNames(processor: Processor<in String>, scope: GlobalSearchScope, filter: IdFilter?) {
+        if (!isEnabled()) return
         StubIndex.getInstance().processAllKeys(indexKey, processor, scope, filter)
     }
 
     override fun processElementsWithName(name: String, processor: Processor<in NavigationItem>, parameters: FindSymbolParameters) {
+        if (!isEnabled()) return
         val project = parameters.project
         val scope = parameters.searchScope
         val idFilter = parameters.idFilter