+++ src/main/kotlin/icu/windea/pls/lang/psi/mock/ParadoxDefinitionNavigationElement.kt
@@ -3,18 +3,18 @@ package icu.windea.pls.lang.psi.mock
 import com.intellij.psi.*
 import icu.windea.pls.*
 import icu.windea.pls.core.*
-import icu.windea.pls.script.psi.ParadoxScriptDefinitionElement
 import icu.windea.pls.model.*
 import icu.windea.pls.script.psi.*
+import java.util.*
 import javax.swing.*
 
 /**
- * 用于在双击shift进行快速查找时显示正确的图标和名字。
+ * 用于在 *随处搜索* 中查找定义。
  */
 class ParadoxDefinitionNavigationElement(
-    element: ParadoxScriptDefinitionElement,
+    val parent: ParadoxScriptDefinitionElement,
     private val definitionInfo: ParadoxDefinitionInfo
-) : ParadoxMockPsiElement(element) {
+) : ParadoxMockPsiElement(parent) {
     override fun getIcon(): Icon {
         return PlsIcons.Nodes.Definition(definitionInfo.type)
     }
@@ -28,14 +28,24 @@ class ParadoxDefinitionNavigationElement(
     }
 
     override fun getText(): String? {
-        return (parent as ParadoxScriptDefinitionElement).text
+        return parent.text
     }
 
     override fun getNameIdentifier(): PsiElement? {
-        return (parent as ParadoxScriptDefinitionElement).nameIdentifier
+        return parent.nameIdentifier
     }
 
     override fun getNavigationElement(): PsiElement {
         return parent
+    }
+
+    override fun equals(other: Any?): Boolean {
+        return this === other || other is ParadoxDefinitionNavigationElement
+            && definitionInfo == other.definitionInfo
+            && project == other.project
+    }
+
+    override fun hashCode(): Int {
+        return Objects.hash(definitionInfo, project)
     }
 }