+++ src/main/kotlin/icu/windea/pls/model/ParadoxDefinitionInfo.kt
@@ -10,7 +10,6 @@ import icu.windea.pls.core.annotations.*
 import icu.windea.pls.core.collections.*
 import icu.windea.pls.core.util.*
 import icu.windea.pls.ep.configContext.*
-import icu.windea.pls.script.psi.ParadoxScriptDefinitionElement
 import icu.windea.pls.lang.util.*
 import icu.windea.pls.script.psi.*
 import java.util.*